(#114) Fix changing `max_waiting` value when `delivery_subject` is set
Here we update the `max_waiting` consumer resource field to be computed.
This allows us to set its value to 0 when `delivery_subject` is set
and not have it flip to the default on every apply.

`max_waiting` now also conficts with the `delivery_subject` field, since
these two imply the use of push vs pull consumers and don't make sense
when specified together.